+++ packages/ui/src/components/common/table/Table.module.css
@@ -55,13 +55,11 @@
 
 @media only screen and (max-width: 564px) {
     .custom-table {
-        font-size: 0.8rem;
         margin: 0.8rem;
     }
 
     .table-header,
     .table-cell {
         padding: 0.4rem 0.8rem;
-        font-size: 0.8rem;
     }
 }
