+++ packages/ui/src/page/documentation/Documentation.jsx
@@ -1,145 +1,10 @@
-import { useState } from "react";
 import styles from "./Documentation.module.css";
-import pythonLogo from "../../assets/python.png";
 import { Link } from "react-router-dom";
-import jsLogo from "../../assets/js.png";
-import javaLogo from "../../assets/java.png";
-import copycodeicon from "../../assets/copy-code-icon.png";
-import tick from "../../assets/tick.png";
 import Table from "../../components/common/table/Table.jsx";
-import PropTypes from "prop-types";
-
-const codeExamples = {
-  javascript: `// Initialize API request to fetch company logo
-  fetch('/api/business/logo?domain={domain}&API_KEY={api_key}', {
-    method: 'GET',
-    headers: {
-      'Content-Type': 'application/json',  // Set content type for JSON response
-    },
-  })`,
-
-  python: `# Import requests library for making HTTP requests
-  import requests
-  
-# Send GET request to fetch company logo
-  response = requests.get('api/business/logo',
-    params={
-        'domain': 'domain',    # Company domain name
-        'API_KEY': 'api_key'   # Your API authentication key
-    },
-    headers={
-        'Content-Type': 'application/json'  # Set content type for JSON response
-    }
-  )`,
-
-  java: `// Create HTTP client instance
-  HttpClient client = HttpClient.newHttpClient();
-  
-  // Build HTTP request with required parameters
-  HttpRequest request = HttpRequest.newBuilder()
-    .uri(URI.create("/api/business/logo?domain=domain&API_KEY=api_key"))
-    .header("Content-Type", "application/json")  // Set content type for JSON response
-    .GET()
-    .build();  
-  
-  // Send request and get response as String
-  HttpResponse<String> response = client.send(request, 
-    HttpResponse.BodyHandlers.ofString());`,
-};
-
-const CodeBlock = ({ id, codeExamples }) => {
-  const [selectedLanguage, setSelectedLanguage] = useState("javascript");
-  const [copyMessage, setCopyMessage] = useState("");
-
-  const copyToClipboard = (code) => {
-    navigator.clipboard
-      .writeText(code)
-      .then(() => {
-        setCopyMessage("Copied!");
-        setTimeout(() => setCopyMessage(""), 2000);
-      })
-      .catch(() => {
-        setCopyMessage("Copy failed!");
-        setTimeout(() => setCopyMessage(""), 2000);
-      });
-  };
-
-  const logos = {
-    javascript: jsLogo,
-    python: pythonLogo,
-    java: javaLogo,
-  };
-
-  return (
-    <div className={styles.codeBlockWrapper}>
-      <div className={styles.codeBlockHeader}>
-        <div className={styles.languageSelector}>
-          {Object.keys(codeExamples).map((lang) => (
-            <button
-              key={`${id}-${lang}`}
-              onClick={() => setSelectedLanguage(lang)}
-              className={`${styles.languageButton} ${
-                selectedLanguage === lang ? styles.languageButtonActive : ""
-              }`}
-            >
-              <img
-                src={logos[lang]}
-                alt={`${lang} logo`}
-                style={{ width: "32px", height: "32px" }}
-              />
-            </button>
-          ))}
-        </div>
-      </div>
-      <pre className={styles.codeBlock}>
-        <div className={styles.copyinside}>
-          <code>{codeExamples[selectedLanguage]}</code>
-          <button
-            onClick={() => copyToClipboard(codeExamples[selectedLanguage])}
-          >
-            {copyMessage ? (
-              <img
-                src={tick}
-                alt="tick"
-                style={{ width: "24px", height: "24px" }}
-              />
-            ) : (
-              <img
-                src={copycodeicon}
-                alt="Copy code"
-                style={{ width: "24px", height: "24px", cursor: "pointer" }}
-              />
-            )}
-          </button>
-        </div>
-      </pre>
-    </div>
-  );
-};
-
-CodeBlock.propTypes = {
-  id: PropTypes.string.isRequired,
-  codeExamples: PropTypes.objectOf(PropTypes.string).isRequired,
-};
+import { codeExamples, tableData } from "../../utils/Constants.js";
+import CodeBlock from "../../components/CodeBlock.jsx";
 
 const Documentation = () => {
-  const tableData = {
-    headers: ["Parameter", "Type", "Description", "Required"],
-    logoRows: [
-      ["domain", "string", "The domain name of the company.", "Yes"],
-      ["API_KEY", "string", "The API Key generated from the dashboard.", "Yes"],
-    ],
-    searchRows: [
-      [
-        "domainKey",
-        "string",
-        "The starting prefix of the domain name to filter logo URLs.",
-        "Yes",
-      ],
-      ["API_KEY", "string", "The API Key generated from the dashboard.", "Yes"],
-    ],
-  };
-
   return (
     <div className="container">
       <div className={styles.card}>
